api-gateway/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==3.0.1
psycopg2-binary==2.9.1
pytest==6.2.4
Werkzeug==3.0.0
flask_cors==3.0.10

requests~=2.32.3
authlib~=0.15.4